backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
google-generativeai
python-multipart
PyMuPDF
Pillow
pytesseract
packaging
numpy
pandas
scikit-learn